refactor(web): move leaderboard data logic into hook
